// coastwatch-web/src/generated/port_intel.ts
export type GeneratedAt = string;
// ...
export type Ports = PortIntel[];
// ...
export type Provenance = Provenance1[];

export type SchemaVersion = 1;

export interface PortIntelCollection {
  generated_at: GeneratedAt;
  method: Method;
  ports: Ports;
  provenance: Provenance;
  schema_version?: SchemaVersion;
}

export interface Method {
  [k: string]: string;
}

export interface PortIntel {
  caveats: Caveats;
  charm: PortCharm | null;
  chlorophyll: PortChlorophyll | null;
  county: County;
  display_name: DisplayName;
  lat: Lat;
  lon: Lon;
  official_relations: OfficialRelations;
  port_code: PortCode;
  region: Region;
}
// ...
export interface Provenance1 {
  citation?: Citation;
  dataset_id?: DatasetId1;
  institution?: Institution;
  license: License;
  pipeline_run_id: PipelineRunId;
  pipeline_version: PipelineVersion;
  product_version?: ProductVersion;
  request_urls?: RequestUrls;
  retrieved_at: RetrievedAt;
  source_id: SourceId;
  source_name: SourceName;
  source_url: SourceUrl1;
  upstream_metadata?: UpstreamMetadata;
}